requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
Pillow>=9.0.0
shapely>=2.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
